Fixed a convex hull bug
There was a bug in the point pruning section of the algorithm. The
lowest point was not always guaranteed to be in the first position after
sorting the array. This could happen if we had two points with the same
y-coord. The lowest point cannot really have an angle with itself, so
the fix is to give it a negative angle to guarantee it is in the first
position in the sorted array.